pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminalfx"
version = "0.1.0"
description = "Offline terminal-style video effects workstation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Gray" }]
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.10",
  "pillow>=10.3",
  "PyYAML>=6.0.1",
  "textual>=0.61.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12",
]

[project.scripts]
terminalfx = "terminalfx.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
packages = ["terminalfx"]
strict = true
warn_unreachable = true
